==> behavior-sequence-transformer-recommender-with-pytorch-by-movie-rating-prediction copy/10-data-transformation.ipynb <==
# A Transformer-based recommendation system


%load_ext autoreload
%autoreload 2


import os
import math
from zipfile import ZipFile
from urllib.request import urlretrieve
import numpy as np
import pandas as pd
from src import utils
from sklearn.preprocessing import MinMaxScaler

## (1)  Load  data

# urlretrieve("http://files.grouplens.org/datasets/movielens/ml-1m.zip", "movielens.zip")
# ZipFile("movielens.zip", "r").extractall()

Then, we load the data into pandas DataFrames with their proper column names.

users = pd.read_csv(
    "ml-1m/users.dat",
    sep="::",
    names=["user_id", "sex", "age_group", "occupation", "zip_code"],
    engine='python'
)

ratings = pd.read_csv(
    "ml-1m/ratings.dat",
    sep="::",
    names=["user_id", "movie_id", "rating", "unix_timestamp"],
    engine='python'
)

movies = pd.read_csv(
    "ml-1m/movies.dat", sep="::", names=["movie_id", "title", "genres"],
    engine='python',encoding='ISO-8859-1'
)

## (2) Remap ID

#### User

users = users.drop("zip_code",axis=1)
users.head()

def generate_remap_id_dict(df,col):
    ids = df[df[col].notnull()][col].unique().tolist()
    ids = sorted(ids)
    id_map_dict = {x: i+1 for i, x in enumerate(ids)}
    id_map_dict["UNK"]=0

    df[f"{col}_index"] = df[col].fillna("UNK").map(id_map_dict)

    return id_map_dict

user_id_map_dict = generate_remap_id_dict(users,col='user_id')
utils.save_object("./artifacts/user_id_map_dict.pkl",user_id_map_dict)

users.head()

sex_id_map_dict = {'M':0.0,'F':1.0,'UNK':0.5}
utils.save_object("./artifacts/sex_id_map_dict.pkl",sex_id_map_dict)
users['sex']=users['sex'].map(sex_id_map_dict)

occupation_id_map_dict = generate_remap_id_dict(users,col='occupation')
utils.save_object("./artifacts/occupation_id_map_dict.pkl",occupation_id_map_dict)
# occupation_id_map_dict

age_group_id_map_dict = generate_remap_id_dict(users,col='age_group')
utils.save_object("./artifacts/age_group_id_map_dict.pkl",age_group_id_map_dict)
age_group_id_map_dict

users.head()

#### Rating

ratings.head()

min_max_scaler = MinMaxScaler()

ratings["norm_rating"] = min_max_scaler.fit_transform(
    ratings["rating"].values.reshape(-1, 1))[:, 0]

utils.save_object("./artifacts/rating_min_max_scaler.pkl",min_max_scaler)

#### Movie

movies.head()

movie_id_id_map_dict = generate_remap_id_dict(movies,col='movie_id')
utils.save_object("./artifacts/movie_id_map_dict.pkl",movie_id_id_map_dict)

genres_set = set()
def get_genres_set(genres):
    global genres_set
    genres_split = genres.split("|")
    genres_set.update(genres_split)
    return genres_split

movies['genres'] = movies['genres'].apply(lambda x:get_genres_set(x))

genres_map_dict ={x: i+1 for i, x in enumerate(sorted(genres_set))}
genres_map_dict['UNK']=0

utils.save_object("./artifacts/genres_map_dict.pkl",genres_map_dict)

movies['genres_ids']= movies['genres'].apply(lambda x: [genres_map_dict[g] for g in x])

ratings['movie_id_index'] = ratings['movie_id'].map(movie_id_id_map_dict) 
ratings['user_id_index'] = ratings['user_id'].map(user_id_map_dict) 

df_user_views = ratings[["movie_id_index", "user_id_index", "unix_timestamp", "norm_rating"]] \
    .merge(movies[['movie_id_index', 'genres_ids']],
           on=['movie_id_index'])
#     .merge(users[['user_id_index', 'sex_index', 'occupation_index', 'age_group_index']],
#            on=['user_id_index']) \

df_user_views.head()

## (3) Transorm into sequences

df_agg_ratings = df_user_views.sort_values(by=["unix_timestamp"]).groupby("user_id_index")

ratings_data = pd.DataFrame(
    data={
        "user_id_index": list(df_agg_ratings.groups.keys()),
        "movie_sequence": list(df_agg_ratings.movie_id_index.apply(list)),
        "rating_sequence": list(df_agg_ratings.norm_rating.apply(list)),
        "timestamps": list(df_agg_ratings.unix_timestamp.apply(list)),
    }
)


ratings_data.head()

sequence_length = 4
step_size = 2


def create_sequences(values, window_size, step_size):
    sequences = []
    start_index = 0
    while True:
        end_index = start_index + window_size
        seq = values[start_index:end_index]
        if len(seq) < window_size:
            seq = values[-window_size:]
            if len(seq) == window_size:
                sequences.append(seq)
            break
        sequences.append(seq)
        start_index += step_size
    return sequences


ratings_data.movie_sequence	 = ratings_data.movie_sequence.apply(
    lambda ids: create_sequences(ids, sequence_length, step_size)
)

ratings_data.rating_sequence = ratings_data.rating_sequence.apply(
    lambda ids: create_sequences(ids, sequence_length, step_size)
)

del ratings_data["timestamps"]

ratings_data.head()

ratings_data_movies = ratings_data[["user_id_index", "movie_sequence"]].explode(
    "movie_sequence", ignore_index=True
)
ratings_data_rating = ratings_data[["rating_sequence"]].explode("rating_sequence", ignore_index=True)
ratings_data_transformed = pd.concat([ratings_data_movies, ratings_data_rating], axis=1)

ratings_data_transformed.head()

user_columns = ['user_id_index',
                'sex',
                'occupation_index',
                'age_group_index']

ratings_data_transformed = ratings_data_transformed.merge(
    users[user_columns], on="user_id_index"
)

ratings_data_transformed['sex'] = ratings_data_transformed['sex'].astype(float)

## mask the target rating

ratings_data_transformed['target_movie'] =  ratings_data_transformed['movie_sequence'].apply(lambda x:x[-1])

ratings_data_transformed['target_rating'] =  ratings_data_transformed['rating_sequence'].apply(lambda x:x[-1])

mask_score = 1.0
ratings_data_transformed['rating_sequence'] =  ratings_data_transformed['rating_sequence'].apply(lambda x:x[:-1]+[mask_score])

ratings_data_transformed.head()

random_selection = np.random.rand(len(ratings_data_transformed.index)) <= 0.85
train_data = ratings_data_transformed[random_selection]
test_data = ratings_data_transformed[~random_selection]

len(train_data)

train_data.to_parquet("./artifacts/train_data.parquet")
test_data.to_parquet("./artifacts/test_data.parquet")
