==> behavior-sequence-transformer-recommender-with-pytorch-by-movie-rating-prediction copy/20-dataset-preparation.ipynb <==
# A Transformer-based recommendation system


%load_ext autoreload
%autoreload 2


import pandas as pd
from torch.utils.data import DataLoader

df_train = pd.read_parquet("./artifacts/train_data.parquet")

df_train.head()

item = df_train.iloc[0]

item.to_dict()

%%writefile ./src/dataset.py
import torch
from torch.utils.data import Dataset


class RatingDataset(Dataset):
    def __init__(self, data):
        self.data = data

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):
        
        item_dict = self.data.iloc[index].to_dict()

        dtype_dict = {}
        for k,v in item_dict.items():
            dtype_dict[k]=torch.long
        dtype_dict['rating_sequence']=torch.float32
        dtype_dict['target_rating']=torch.float32
        dtype_dict['sex']=torch.float32


        sample = {}
        for k,v in item_dict.items():
            sample[k] = torch.tensor(v,dtype=dtype_dict[k])
            
        return sample

from src.dataset import RatingDataset

train_dataset = RatingDataset(data=df_train) 

train_dataset[0]

loader = DataLoader(train_dataset,batch_size=4,shuffle=True)

for sample in loader:
    break


sample
