==> behavior-sequence-transformer-recommender-with-pytorch-by-movie-rating-prediction copy/30-modeling.ipynb <==
# A Transformer-based recommendation system


%load_ext autoreload
%autoreload 2


import pandas as pd
from src.dataset import RatingDataset
from src import utils
from torch.utils.data import DataLoader
from torch import nn
import math
import torch

### (1) Load Dataset

df_train = pd.read_parquet("./artifacts/train_data.parquet")
df_test = pd.read_parquet("./artifacts/test_data.parquet")

train_dataset = RatingDataset(data=df_train) 

loader = DataLoader(train_dataset,batch_size=12,shuffle=True)

for inputs in loader:
    break


## (2) Model Config

# inputs

age_group_id_map_dict = utils.open_object("./artifacts/age_group_id_map_dict.pkl")

movie_id_map_dict = utils.open_object("./artifacts/movie_id_map_dict.pkl")

occupation_id_map_dict = utils.open_object("./artifacts/occupation_id_map_dict.pkl")

sex_id_map_dict = utils.open_object("./artifacts/sex_id_map_dict.pkl")

user_id_map_dict = utils.open_object("./artifacts/user_id_map_dict.pkl")
# genres_map_dict = utils.open_object("./artifacts/genres_map_dict.pkl")

rating_min_max_scaler = utils.open_object("./artifacts/rating_min_max_scaler.pkl")

num_user = len(user_id_map_dict)
num_movie = len(movie_id_map_dict)
num_occupation = len(occupation_id_map_dict)
num_age_group = len(age_group_id_map_dict)
# num_genre = len(genres_map_dict)

embed_configs = {}
EMED_DIM=32
sequence_length=4
embed_configs['user']={"embed_dim":EMED_DIM,"num_embed":num_user}
embed_configs['movie']={"embed_dim":EMED_DIM,"num_embed":num_movie}
embed_configs['occupation']={"embed_dim":EMED_DIM,"num_embed":num_occupation}
embed_configs['age_group']={"embed_dim":EMED_DIM,"num_embed":num_age_group}
embed_configs['position'] = {"embed_dim":EMED_DIM,"num_embed":sequence_length}

config_dict={}
config_dict['embed_configs'] = embed_configs
config_dict['transformer_num_layer']=3
config_dict['dropout']=0.2

class Config:
    def __init__(self, dictionary):
        for key, value in dictionary.items():
            setattr(self, key, value)

config = Config(dictionary=config_dict)

config.embed_configs

## Modeling

### (1) Create Embedding Layer

class self:
    pass

self.config = config

embed_configs = self.config.embed_configs

embedding_layers = []
for name,embed_config in embed_configs.items():
    embed_dim = embed_config['embed_dim']
    num_embed = embed_config['num_embed']
    embeding_layer = nn.Embedding(num_embeddings=num_embed, embedding_dim=embed_dim)
    nn.init.xavier_uniform_(embeding_layer.weight)
    embedding_layers.append([name,embeding_layer])

self.embedding_layers = nn.ModuleDict(embedding_layers)

self.embedding_layers

inputs.keys()

### (2) Embed Movie

inputs['target_movie']

target_movie_embedding = self.embedding_layers['movie'](inputs['target_movie'])

target_movie_embedding.shape

movie_sequence_embedding  = self.embedding_layers['movie'](inputs['movie_sequence'])

movie_sequence_embedding.shape

### (4) Position Embedding

self.config.embed_configs['position']

positions = torch.arange(self.config.embed_configs['position']['num_embed'])

position_embedding = self.embedding_layers['position'](positions)

position_embedding.shape

### (5) Sequence Rating

rating_sequence = inputs['rating_sequence']

rating_sequence.shape

inputs.keys()

### (5) Embedding User

self.embedding_layers

sex_input = inputs['sex']
user_embedding = self.embedding_layers['user'](inputs['user_id_index'])
occupation_embedding = self.embedding_layers['occupation'](inputs['occupation_index'])
age_group_embedding = self.embedding_layers['age_group'](inputs['age_group_index'])

### (6) User Behavior Sequence Input

position_embedding.shape

batch_size = movie_sequence_embedding.shape[0]

batch_position_embedding = torch.stack([position_embedding.clone() for _ in range(batch_size)])
batch_position_embedding.shape

movie_sequence_embedding.shape

movie_pos_seq_embedding = torch.concat([movie_sequence_embedding,batch_position_embedding],dim=-1)

movie_pos_seq_embedding.shape

## cross with rating

rating_sequence.unsqueeze(-1).shape

torch.mul(movie_pos_seq_embedding,rating_sequence.unsqueeze(-1)).shape

movie_pos_rating_seq_embedding = torch.mul(movie_pos_seq_embedding,rating_sequence.unsqueeze(-1))

### (6) Transformer for sequence features

%%writefile ./src/transformer_layer.py

import torch
import torch.nn as nn
import torch.nn.functional as F

class TransformerBlock(nn.Module):
    def __init__(self, input_size, output_size, num_heads, dropout_rate):
        super(TransformerBlock, self).__init__()

        self.multihead_attention = nn.MultiheadAttention(input_size, num_heads)
        self.layer_norm1 = nn.LayerNorm(input_size)

        self.feed_forward = nn.Sequential(
            nn.Linear(input_size, 4*input_size),
            nn.ReLU(),
            nn.Linear(4*input_size, output_size),
            nn.Dropout(dropout_rate)
        )
        self.layer_norm2 = nn.LayerNorm(output_size)
        self.dropout = nn.Dropout(dropout_rate)

    def forward(self, x):
        # Multi-head Attention
        attn_output, _ = self.multihead_attention(x, x, x)
        x = self.layer_norm1(x + attn_output)

        # Feed-Forward Network
        ff_output = self.feed_forward(x)
        x = self.layer_norm2(x + ff_output)
        x = self.dropout(x)
        return x

class TransformerLayer(nn.Module):
    def __init__(self, d_model, num_heads=8, dropout_rate=0.2, num_layers=3):
        super(TransformerLayer, self).__init__()

        self.transformer_blocks = nn.ModuleList([
            TransformerBlock(d_model, d_model, num_heads, dropout_rate)
            for _ in range(num_layers)
        ])

    def forward(self, x):
        for transformer_block in self.transformer_blocks:
            x = transformer_block(x)
        return x


src = torch.rand(10, 32, 64)

from src.transformer_layer import TransformerLayer

transformer_layer = TransformerLayer(d_model=64,num_heads=8, dropout_rate=0.2, num_layers=3)

transformer_layer(src).shape

movie_pos_rating_seq_embedding.shape

seq_transformer_output = transformer_layer(movie_pos_rating_seq_embedding)

seq_transformer_output.shape

seq_transformer_flatten_output = seq_transformer_output.view(12,-1)

seq_transformer_flatten_output.shape

### (7) Concat other Features

inputs.keys()

sex_feature = inputs['sex']
user_embedding = self.embedding_layers['user'](inputs['user_id_index'])
occupation_embedding = self.embedding_layers['occupation'](inputs['occupation_index'])
age_group_embedding = self.embedding_layers['age_group'](inputs['age_group_index'])

target_movie_embedding.shape

sex_feature = sex_feature.unsqueeze(-1)
sex_feature.shape

sex_cross_feature = torch.mul(sex_feature,target_movie_embedding)

user_embedding.shape

user_embedding_cross = torch.mul(user_embedding,target_movie_embedding)

occupation_embedding_cross = torch.mul(occupation_embedding,target_movie_embedding,)

age_group_embedding_cross = torch.mul(age_group_embedding,target_movie_embedding)

user_features = torch.concat([sex_feature,user_embedding,occupation_embedding,age_group_embedding],dim=-1)

user_cross_features = torch.concat([sex_cross_feature,user_embedding_cross,occupation_embedding_cross,age_group_embedding_cross],dim=-1)

user_inputs_features = torch.concat([user_features,user_cross_features,target_movie_embedding],axis=1)

user_inputs_features.shape

#### MLP Layer 

%%writefile ./src/mlp_layer.py

import torch.nn as nn

class MLP(nn.Module):
    def __init__(self, dropout=0.2, hidden_units=[512, 256,128]):
        super(MLP, self).__init__()
        self.dropout = nn.Dropout(p=dropout)
        self.layers = nn.ModuleList()
        for i in range(len(hidden_units) - 1):
            self.layers.append(nn.Linear(hidden_units[i], hidden_units[i + 1]))
            self.layers.append(nn.LeakyReLU())
            self.layers.append(nn.Dropout(p=dropout))
        self.fc = nn.Linear(hidden_units[-1],1)
        self.sigmoid = nn.Sigmoid()
        
    def forward(self, x):
        for layer in self.layers:
            x = layer(x)
        logits = self.fc(x)
        output = self.sigmoid(logits)
        return output


mlp_input_features = torch.concat([user_inputs_features,seq_transformer_flatten_output],axis=1)

mlp_input_features.shape

from src.mlp_layer import MLP

mlp = MLP(dropout=0.2, hidden_units=[513, 256,128])

mlp

outputs = mlp(mlp_input_features)

outputs

loss_func = torch.nn.BCELoss()

loss_func(outputs,inputs['target_rating'].view(-1,1))
